// client/src/api/models/RecurringNewPayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FinchAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    TimeInterval,
    TimeIntervalFromJSON,
    TimeIntervalFromJSONTyped,
    TimeIntervalToJSON,
} from './';

/**
 * 
 * @export
 * @interface RecurringNewPayload
 */
export interface RecurringNewPayload {
    /**
     * 
     * @type {string}
     * @memberof RecurringNewPayload
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof RecurringNewPayload
     */
    start: number;
    /**
     * 
     * @type {number}
     * @memberof RecurringNewPayload
     */
    end: number;
    /**
     * 
     * @type {number}
     * @memberof RecurringNewPayload
     */
    principal: number;
    /**
     * 
     * @type {number}
     * @memberof RecurringNewPayload
     */
    amount: number;
    /**
     * 
     * @type {number}
     * @memberof RecurringNewPayload
     */
    interest: number;
    /**
     * 
     * @type {TimeInterval}
     * @memberof RecurringNewPayload
     */
    frequency: TimeInterval;
}

export function RecurringNewPayloadFromJSON(json: any): RecurringNewPayload {
    return RecurringNewPayloadFromJSONTyped(json, false);
}

export function RecurringNewPayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecurringNewPayload {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'start': json['start'],
        'end': json['end'],
        'principal': json['principal'],
        'amount': json['amount'],
        'interest': json['interest'],
        'frequency': TimeIntervalFromJSON(json['frequency']),
    };
}

export function RecurringNewPayloadToJSON(value?: RecurringNewPayload | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'start': value.start,
        'end': value.end,
        'principal': value.principal,
        'amount': value.amount,
        'interest': value.interest,
        'frequency': TimeIntervalToJSON(value.frequency),
    };
}